Fix conversion warnings from CI

+++ Os/Posix/File.cpp
@@ -206,7 +206,10 @@ PosixFile::Status PosixFile::read(U8* buffer, FwSignedSizeType &size, PosixFile:
 
     for (FwSignedSizeType i = 0; i < maximum && accumulated < size; i++) {
         // char* for some posix implementations
-        ssize_t read_size = ::read(this->m_handle.m_file_descriptor, reinterpret_cast<CHAR*>(&buffer[accumulated]), size - accumulated);
+        ssize_t read_size = static_cast<ssize_t>(::read(
+            this->m_handle.m_file_descriptor,
+            reinterpret_cast<CHAR*>(&buffer[accumulated]),
+            size - accumulated));
         // Non-interrupt error
         if (PosixFileHandle::ERROR_RETURN_VALUE == read_size) {
             PlatformIntType errno_store = errno;
